fix: pool snapshot
