+++ schema.graphql
@@ -214,6 +214,7 @@ type PoolSnapshot @entity {
   totalShares: BigDecimal!
   swapVolume: BigDecimal!
   swapFees: BigDecimal!
+  liquidity: BigDecimal!
   timestamp: Int!
 }
